model-training/tracking/requirements.txt
ultralytics>=8.2.100
torch>=2.1.0
torchvision>=0.16.0
opencv-python>=4.9.0
numpy>=1.26
pandas>=2.2
pyyaml>=6.0.1
tqdm>=4.66
rich>=13.7
scipy>=1.11
filterpy>=1.4.5
lapx>=0.5.5
matplotlib>=3.8
scikit-learn>=1.3
motmetrics>=1.4
